OTP email send successful
