+++ api/authenticate.js
@@ -5,6 +5,7 @@ const otpGenerator = require('otp-generator');
 const Token = require('../models/Token');
 const User = require('../models/User');
 const OTP = require('../models/OTP');
+const nodemailer = require('nodemailer');
 
 /*
 Functionality: Initializes the Fitbit API client
@@ -51,10 +52,14 @@ router.get('/authorize', async(req, res) => {
     `);
 });
 
+// Processes the email based on typr of user
 router.post('/process-email', async(req, res) => {
     const { email, user_type } = req.body;
     const existingUser = await User.findOne({ email: email, user_type: user_type });
+
+    // New User
     if (!existingUser) {
+    // Change when connected to frontend
       return res.send(`Create New User
         <form action="/authenticate/new-user" method="post">
             Email: <input type="text" name="email" value="${email}" readonly><br>
@@ -65,21 +70,52 @@ router.post('/process-email', async(req, res) => {
         </form>
         `);
     }
+
+    // Fitbit User
     console.log("Existing User:");
     if (existingUser.user_type == "FitBit") {
         console.log("FitBit");
         return res.redirect('./authorize-fitbit');
     }
+
+    // Manual User
     console.log("Manual");
     // Sends a one-time password and awaits authentication
     try {
         let otp;
+        // Generates an otp
         do {
             otp = otpGenerator.generate(8, { specialChars: false });
-        } while (await OTP.findOne({ otp : otp}));
-        await OTP.create({ email, otp });
+        } while (await OTP.findOne({ otp : otp }));
 
-        return res.send(`OTP send successfully
+        // Adds otp to database
+        await OTP.create({ email, otp });
+        console.log("New OTP saved");
+
+        // Create a Transporter to send emails
+        let transporter = nodemailer.createTransport({
+            service: 'gmail', // Need proper authentication in order to send email
+            auth: {
+                user: process.env.EMAIL_SENDER,
+                pass: process.env.EMAIL_PASS
+            }
+        });
+        console.log("Transporter created");
+
+        // Send emails to users
+        await transporter.sendMail({
+            from: process.env.EMAIL_SENDER,
+            to: email,
+            subject: "OTP Verification Email", // We can change subject and message as well
+            html: `<h1>Please confirm your OTP</h1>
+                   <p>Here is your OTP for verification: ${otp}</p>
+                   <br>
+                   <em>Do not reply to this email. This email address is being monitered.</em>`,
+        });
+        console.log("Email sent successfully");
+
+        // Change when connected to frontend
+        return res.send(`OTP sent successfully
             <form action="/authenticate/authorize-otp" method="post">
                 Email: <input type="text" name="email" value="${email}" readonly><br>
                 OTP: <input type="text" name="otp" required><br>
@@ -91,7 +127,7 @@ router.post('/process-email', async(req, res) => {
     }
 });
 
-// 
+// Creates a new user and adds them to the database
 router.post('/new-user', async(req, res) => {
     try {
         const { email, user_type } = req.body;
@@ -134,7 +170,7 @@ router.get('/authorize-fitbit', async (req, res) => {
 });
 
 // Checks if OTP is valid and authenticates if it is
-router.get('/authorize-otp', async(req, res) => {
+router.post('/authorize-otp', async(req, res) => {
     const { email, otp } = req.body;
     try {
         const otpRecord = await OTP.findOne({ email, otp });
